requirements.txt
python-telegram-bot==21.3
Pillow==10.3.0
requests==2.32.3
aiohttp==3.9.5
python-docx==1.1.0

